refactor(bridge): separate graphql request logics
