+++ bridge/src/interfaces/headless-graphql-client.ts
@@ -1,0 +1,6 @@
+import { TxId } from "../types/txid";
+
+export interface IHeadlessGraphQLClient {
+    getNextTxNonce(address: string): Promise<number>;
+    transfer(recipient: string, amount: string, txNonce: number): Promise<TxId>;
+}
