+++ bridge/src/ncg-transfer.ts
@@ -1,59 +1,22 @@
-import axios, { AxiosResponse } from "axios";
+import { IHeadlessGraphQLClient } from "./interfaces/headless-graphql-client";
 import { INCGTransfer } from "./interfaces/ncg-transfer";
 import { TxId } from "./types/txid";
 
-interface GraphQLRequestBody {
-    operationName: string | null;
-    query: string;
-    variables: Object;
-}
-
 export class NCGTransfer implements INCGTransfer {
-    private readonly _apiEndpoint: string;
+    private readonly _headlessGraphQLCLient: IHeadlessGraphQLClient;
 
     /**
      * The address of sender;
      */
     private readonly _address: string;
 
-    constructor(apiEndpoint: string, address: string) {
-        this._apiEndpoint = apiEndpoint;
+    constructor(headlessGraphQLCLient: IHeadlessGraphQLClient, address: string) {
+        this._headlessGraphQLCLient = headlessGraphQLCLient;
         this._address = address;
     }
 
     async transfer(address: string, amount: BigInt): Promise<TxId> {
-        const nextTxNonce = await this.getNextTxNonce(this._address);
-        const query = `mutation TransferGold($recipient: Address!, $amount: String!, $txNonce: Long!)
-        { transfer(recipient: $recipient, amount: $amount, txNonce: $txNonce) }`;
-        const response = await this.graphqlRequest({
-            operationName: "TransferGold",
-            query,
-            variables: {
-                recipient: address,
-                amount: amount.toString(),
-                txNonce: nextTxNonce,
-            }});
-
-        return response.data.data.transfer;
-    }
-
-    private async getNextTxNonce(address: string): Promise<number> {
-        const query = "query GetNextTxNonce($address: Address!) { nextTxNonce(address: $address) } ";
-        const response = await this.graphqlRequest({
-            operationName: "GetNextTxNonce",
-            query,
-            variables: { address, }
-        });
-
-        return response.data.data.nextTxNonce;
-    }
-
-    private async graphqlRequest(body: GraphQLRequestBody): Promise<AxiosResponse> {
-        return axios.post(this._apiEndpoint, body,
-        {
-            headers: {
-                "Content-Type": "application/json",
-            },
-        });
+        const nextTxNonce = await this._headlessGraphQLCLient.getNextTxNonce(this._address);
+        return await this._headlessGraphQLCLient.transfer(address, amount.toString(), nextTxNonce);
     }
 }
